# apps/api/contracts/invite_binding.openapi.yaml
openapi: 3.1.0
info:
  title: GameMulti Invite & Binding API Draft
  version: 0.1.0
  description: |
    邀请制账户与绑定流程接口草稿，覆盖邀请码校验、邀请注册、插件发起绑定会话、按 token / pair code 查询会话、确认绑定。
servers:
  - url: https://api.example.com
paths:
  /api/invitations/validate:
    post:
      summary: 校验邀请码是否可用
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        '200':
          description: 校验结果
          content:
            application/json:
              schema:
                type: object
                required: [valid, codeStatus]
                properties:
                  valid:
                    type: boolean
                  codeStatus:
                    type: string
                    enum: [active, disabled, expired, exhausted]
                  remainingUses:
                    type: integer
                  expiresAt:
                    type: string
                    format: date-time
                  message:
                    type: string
  /api/auth/invite-register:
    post:
      summary: 通过邀请码注册账户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, inviteCode]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 32
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                inviteCode:
                  type: string
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '409':
          description: 用户名/邮箱冲突或邀请码不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/plugin/bindings/session:
    post:
      summary: 插件发起绑定会话
      security:
        - pluginClientKey: []
          pluginTimestamp: []
          pluginNonce: []
          pluginSignature: []
          pluginProtocolVersion: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestId, serverCode, gameCode, platform, gameUserId, bindMode]
              properties:
                requestId:
                  type: string
                  minLength: 8
                  maxLength: 128
                  description: Stable per-attempt idempotency key. Retries use the same value with a new HMAC nonce.
                serverCode:
                  type: string
                gameCode:
                  type: string
                platform:
                  type: string
                gameUserId:
                  type: string
                displayName:
                  type: string
                bindMode:
                  type: string
                  enum: [register_new, bind_existing]
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindingSessionCreated'
        '401':
          description: 插件凭证或签名失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginErrorResponse'
        '403':
          description: 服务器待审核或已阻塞
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginErrorResponse'
        '409':
          description: nonce 重放或 requestId 与原请求不匹配
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginErrorResponse'
        '426':
          description: 插件协议不受支持
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginErrorResponse'
  /api/plugin/bindings/{sessionId}:
    get:
      summary: 插件按会话 ID 查询自身服务器的绑定状态
      security:
        - pluginClientKey: []
          pluginTimestamp: []
          pluginNonce: []
          pluginSignature: []
          pluginProtocolVersion: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 不含 binding token 的绑定状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginBindingSessionView'
        '404':
          description: 会话不存在或不属于当前插件客户端/服务器
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginErrorResponse'
        '401':
          description: 插件凭证或签名失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginErrorResponse'
        '403':
          description: 服务器待审核或已阻塞
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginErrorResponse'
        '426':
          description: 插件协议不受支持
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginErrorResponse'
  /api/bindings/session/by-token:
    get:
      summary: 根据 token 查询绑定会话
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 会话详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindingSessionView'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bindings/session/by-pair-code:
    post:
      summary: 根据配对码查询绑定会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pairCode]
              properties:
                pairCode:
                  type: string
                  minLength: 6
                  maxLength: 6
      responses:
        '200':
          description: 会话详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindingSessionView'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bindings/confirm:
    post:
      summary: 确认绑定
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId:
                  type: string
      responses:
        '200':
          description: 绑定成功
          content:
            application/json:
              schema:
                type: object
                required: [bindingId, gameAccountId, status]
                properties:
                  bindingId:
                    type: string
                  gameAccountId:
                    type: string
                  status:
                    type: string
                    enum: [active]
        '409':
          description: 会话已用、过期或账号冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    pluginClientKey:
      type: apiKey
      in: header
      name: X-GM-Client-Key
    pluginTimestamp:
      type: apiKey
      in: header
      name: X-GM-Timestamp
    pluginNonce:
      type: apiKey
      in: header
      name: X-GM-Nonce
    pluginSignature:
      type: apiKey
      in: header
      name: X-GM-Signature
    pluginProtocolVersion:
      type: apiKey
      in: header
      name: X-GM-Protocol-Version
  schemas:
    UserProfile:
      type: object
      required: [id, username, email, status]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, disabled, banned]
    BindingSessionCreated:
      type: object
      required: [sessionId, token, pairCode, expiresIn, bindUrl, publicBindUrl]
      properties:
        sessionId:
          type: string
        token:
          type: string
        pairCode:
          type: string
        expiresIn:
          type: integer
        bindUrl:
          type: string
        publicBindUrl:
          type: string
          format: uri
    BindingSessionView:
      allOf:
        - $ref: '#/components/schemas/BindingSessionDetails'
        - type: object
          required: [id]
          properties:
            id:
              type: string
    PluginBindingSessionView:
      allOf:
        - $ref: '#/components/schemas/BindingSessionDetails'
        - type: object
          required: [sessionId]
          properties:
            sessionId:
              type: string
    BindingSessionDetails:
      type: object
      required: [game, server, platform, gameUserId, status, sessionStatus, recoverable, nextAction, expiresAt, expired]
      properties:
        game:
          type: object
          required: [code, name]
          properties:
            code:
              type: string
            name:
              type: string
        server:
          type: object
          required: [serverCode, serverName]
          properties:
            serverCode:
              type: string
            serverName:
              type: string
        platform:
          type: string
        gameUserId:
          type: string
        displayName:
          type: string
          nullable: true
        bindMode:
          type: string
        status:
          type: string
          enum: [pending, bound, expired, cancelled, conflict, revoked, denied, unavailable]
        sessionStatus:
          type: string
          enum: [pending, confirmed, expired, cancelled, conflict, revoked, denied, unavailable]
        recoverable:
          type: boolean
        nextAction:
          type: string
          enum: [authenticate_with_discourse, confirm_binding, enter_game_or_community, return_to_source, start_new_binding, contact_operations_or_unbind, rebind_or_view_reason, view_authorization_requirements, retry_later]
        expiresAt:
          type: string
          format: date-time
        expired:
          type: boolean
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    PluginErrorResponse:
      type: object
      required: [code, message, retryable, requestId]
      properties:
        code:
          type: string
          enum: [AUTHENTICATION_FAILED, CLOCK_SKEW, NONCE_REPLAY, SERVER_PENDING_APPROVAL, SERVER_BLOCKED, RATE_LIMITED, INVALID_REQUEST, PROTOCOL_UNSUPPORTED, SERVICE_UNAVAILABLE, IDEMPOTENCY_CONFLICT]
        message:
          type: string
        retryable:
          type: boolean
        requestId:
          type: string
        serverTime:
          type: integer
          description: Unix epoch seconds; present for CLOCK_SKEW.
        details:
          type: object
          additionalProperties: true